requirements.txt
python-docx==0.8.11
papersize==1.3.0
docx2pdf==0.1.8